backend/requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.7.0
django-filter==25.1
django-tailwind==3.8.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
joblib==1.4.2
mysqlclient==2.2.7
numpy==2.2.4
pandas==2.2.3
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
scikit-learn==1.6.1
scipy==1.15.2
setuptools==78.1.0
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.1
wheel==0.45.1
